requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.11
Pillow==10.4.0
python-dotenv==1.0.0
cloudinary==1.44.2
django-cloudinary-storage==0.3.0
openai==1.77.0
requests==2.32.5
